cli: add 'network logs' command
